requirements.txt
alembic==1.0.10
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
kiwisolver==1.1.0
Mako==1.0.12
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.1.0
nltk==3.4.4
numpy==1.16.4
packaging==19.0
pandas==0.25.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
SQLAlchemy==1.3.4
textblob==0.15.3
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
